backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
edge-tts>=6.1.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
